# docs/swagger.yaml
basePath: /
definitions:
  models.AddConditionReq:
    properties:
      condition:
        description: |-
          The condition to be added.
          required: true
          min length: 2
          max length: 30
        maxLength: 30
        minLength: 2
        type: string
    required:
    - condition
    type: object
...
  models.DiagnosesReq:
    properties:
      name:
        description: |-
          Name represents the name of the diagnosis.
          required: true
          min length: 2
          max length: 30
        maxLength: 30
        minLength: 2
        type: string
    required:
    - name
    type: object
  models.FailureResponse:
    description: Standard error response format with status and error message.
    properties:
      error:
        example: Bad Request
        type: string
      status:
        example: 400
        type: integer
    type: object
  models.RegAllergyReq:
    description: A request to register a new allergy for a patient
    properties:
      name:
        description: |-
          @Param name query string true "Allergy Name" validate:"required,min=2,max=100"
          @example "Peanut"
        maxLength: 100
        minLength: 2
        type: string
      reaction:
        description: |-
          @Param reaction query string true "Reaction to the allergy" validate:"required,min=2,max=255"
          @example "Swelling"
        maxLength: 255
        minLength: 2
        type: string
      severity:
        description: |-
          @Param severity query string true "Severity of the allergy" validate:"required,oneof=mild moderate severe"
          @example "mild"
        enum:
        - mild
        - moderate
        - severe
        type: string
    required:
    - name
    - reaction
    - severity
    type: object
...
  models.SuccessResponse:
    description: Standard success response format with an optional data field.
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      status:
        example: 200
        type: integer
    type: object
  models.UpdateAllergyReq:
    description: A request to update an existing allergy record
    properties:
      name:
        description: |-
          @example "Peanut"
          @Param name query string false "Updated name of the allergy" validate:"omitempty,min=2,max=100"
        maxLength: 100
        minLength: 2
        type: string
      reaction:
        description: |-
          @example "swelling"
          @Param reaction query string false "Updated reaction to the allergy" validate:"omitempty,min=2,max=255"
        maxLength: 255
        minLength: 2
        type: string
      severity:
        description: |-
          @example "moderate"
          @Param severity query string false "Updated severity of the allergy" validate:"omitempty,oneof=mild moderate severe"
        enum:
        - mild
        - moderate
        - severe
        type: string
    type: object
  models.UpdateDiagnosesReq:
    properties:
      name:
        description: |-
          Name represents the updated name of the diagnosis.
          required: true
          min length: 2
          max length: 30
        maxLength: 30
        minLength: 2
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: codervaidik@gmail.com
    name: Vaidik Bajpai
  description: Backend API for MediBridge, a medical record management system.
  title: MediBridge API
  version: "1.0"
paths:
  /v1/allergy/{allergyID}:
    delete:
      consumes:
      - application/json
      description: Deletes an allergy from the patient’s record by its ID.
      parameters:
      - description: Allergy ID (UUID)
        in: path
        name: allergyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Delete an allergy
      tags:
      - Allergy
    put:
      consumes:
      - application/json
      description: Updates an existing allergy using its ID.
      parameters:
      - description: Allergy ID (UUID)
        in: path
        name: allergyID
        required: true
        type: string
      - description: Updated allergy details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAllergyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Update an allergy
      tags:
      - Allergy
  /v1/condition/{conditionID}:
    delete:
      consumes:
      - application/json
      description: Marks an existing condition as inactive by its ID.
      parameters:
      - description: Condition ID (UUID)
        in: path
        name: conditionID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Inactivate a medical condition
      tags:
      - Conditions
  /v1/diagnoses/{diagnosesID}:
    delete:
      consumes:
      - application/json
      description: Deletes a diagnosis using the diagnosis ID.
      parameters:
      - description: Diagnosis ID (UUID)
        in: path
        name: diagnosesID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Delete a diagnosis
      tags:
      - Diagnoses
    put:
      consumes:
      - application/json
      description: Updates an existing diagnosis using the diagnosis ID.
      parameters:
      - description: Diagnosis ID (UUID)
        in: path
        name: diagnosesID
        required: true
        type: string
      - description: Updated diagnosis details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDiagnosesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Update an existing diagnosis
      tags:
      - Diagnoses
...
  /v1/patient/{patientID}/allergy:
    post:
      consumes:
      - application/json
      description: Records a new allergy for the specified patient.
      parameters:
      - description: Patient ID (UUID)
        in: path
        name: patientID
        required: true
        type: string
      - description: Allergy input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RegAllergyReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Record a new allergy
      tags:
      - Allergy
  /v1/patient/{patientID}/condition:
    post:
      consumes:
      - application/json
      description: Adds a new medical condition associated with a patient ID.
      parameters:
      - description: Patient ID (UUID)
        in: path
        name: patientID
        required: true
        type: string
      - description: Condition details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AddConditionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Add a new medical condition
      tags:
      - Conditions
  /v1/patient/{patientID}/diagnoses:
    post:
      consumes:
      - application/json
      description: Adds a new diagnosis for a patient using their patient ID.
      parameters:
      - description: Patient ID (UUID)
        in: path
        name: patientID
        required: true
        type: string
      - description: Diagnosis details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.DiagnosesReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Add a new diagnosis
      tags:
      - Diagnoses
...
schemes:
- http
swagger: "2.0"